Game now functional, but needs playtesting

+++ RestaurantSimulator/Models/Ingrediants.cs
@@ -3,7 +3,7 @@ namespace RestaurantSimulator.Models
     public class Ingredients
     {
         public required string Name { get; set; }
-        public required int InitialStock { get; set; }
+        public required double InitialStock { get; set; }
         public required string Unit { get; set; }
         public required double Cost { get; set; }
         public string Status { get; set; } = string.Empty;
